community/mpv: compile with rsound support
This adds RSound support to mpv using librsound. This allows you to
use the audio output from another computer on your network. This feature
is similar to PluseAudios TCP module.

Sending this to patchwork instead of committing it myself because it is
rather uncommon to compile mpv with rsound support by default, however,
it doesn't hurt and I find this feature very useful.