main/openrc: refactor modloop to not depend on alpine_dev boot opt
We scan all relevant mount points for the modloop specified with the
modloop boot option. If modloop boot option is missing, then we warn
about it, look for files that blkid recognizes, try mount them and then
look for modules/$(uname -r).

Wen unmounting, find the loop device, and the backing_file and the
mountpoint that holds the backing file.