testing/editorconfig: new aport
EditorConfig helps developers define and maintain consistent
coding styles between different editors and IDEs.

This package provides the standalone editorconfig core binary
and library that is needed by various plugins.

For example, the vim plugin (not in this package)
depends on the editorconfig binary (this package)
since vim is compiled without python support on Alpine.