main/ulogd: Added description to init.d script and removed net dependency
Added description to init.d script in order to make it compliant
with rc-system and avoid annoying warning messages on syslog.
Deleted net dependency in order to avoid the error on startup
"cannot start ulogd as networking would not start", since the
required before firewall and firewall start before net.