testing/mongodb: Fix mongo shell "too much recursion" #5117
The problem is due to the fact that spidermonkey expects a
larger available space on the thread stack than the one
allocated by default by Musl libc.

In the code which checks the available stack space, we adjust
the reserved quota to 32k instead of 64k because the default
musl libc thread stack size is small (80k) and we can not set
stack size for C++ threads.
Note also that with musl, stack size does not include guard page.