testing/radare2: upgrade to 0.10.1 and fix license
The documentation on licensing in radare2 says there's a lot of GPL
plugins which we are distributing, the rest is LGPL and given the
virulent nature of the GPL they become GPL when it comes to distributing
them. So the safest thing to do is just leaving it as GPLv3.