main/busybox-initscripts: rename cron service to crond
cron is the name of a virtual service and a real service. This causes a
non-fatal error with the most recent OpenRC version. This error appears
everytime openrc updates the dependency tree this can be triggered
manually by calling rc_deptree_update(3).

Renaming the busybox cron service fixes this error, however, this breaks
backwards compatibility.