Revert "main/perl-package-deprecationmanager: upgrade to 0.16"
The upgrade introduced a new dependency (perl-namespace-autoclean) which
is currently still in testing/ and since packages in main/ cannot depend
on package in testing/ this breaks shit.

Moving perl-namespace-autoclean to main/ doesn't solve the problem
entirely because it depends on perl-moose which than again depends on
perl-package-deprecationmanager therefore we would have a circular
dependency which needs to be resolved somehow.

This reverts commit 6562cd444773e9c8d6e75d9bba7fb315d19f6a86.