main/rsyslog: update to 8.18.0, and apply a new musl fix
This commit brings rsyslog up to the latest release that has some nice performance improvements and bug fixes.
This commit also includes an upstream fix from 8.19.0 (unreleased) which fixes a segfault present in all current rsyslog versions on musl.
Version 8.17 and later of rsyslog has a new dependency. It now needs libfastjson installed to run.
I created the libfastjson package earlier this month, it is in the `testing` directory. I believe libfastjson will need to be promoted from `testing` to `edge`.