main/lighttpd: Fix lighttpd user's primary group.
During mass .pre-install/upgrade init scripts changes there was a change
not preserving status quo regarding group creation and user creation for
57 packages that weren't using -S option with adduser calls earlier.
It led to not creating the group named like the user and changing
primary user group from the same named to nogroup.

It has been noticed thanks to lighttpd issue (refs #5666) and it has
been partially fixed for lighttpd alone already. This patch completes
the fix and is also an announcement for some mass changes coming later.