main/gcc: upgrade to gcc 6 snapshot 20160721
this includes additional arm fixes that fix u-boot compilation.

_pkgbase is expected to be the release tarball, and we have
cumulative diff to bring up to the snapshot.

gcc marks the snapshot releases as 6.1.1 too, so we use this
version for our packaging too. gcc -v reports:
  gcc version 6.1.1 20160721 (Alpine 6.1.1)

so the actual snapshot date is visible too. next release is
due to be 6.2.