testing/llvm3.7: move libLLVM.so to /usr/lib, add version to libs name
This is needed for binaries that are dynamically linked with
libLLVM, so they can find it on the default LD_LIBRARY_PATH.

To be honest, I'm still not sure if everything is right. LLVM's CMake
scripts have a lot flaws in this version (see recent commits). Most
distribution still use old make files.

However, it already works for Rust (linked statically) and Julia (linked
dynamically).