main/quagga: upgrade to 1.1.0, two CVEs
CVE-2016-1245 Stack buffer overflow in zebra on Linux, if IPv6  and
IPv6 neighbour discovery router advertisements (SLAC) are enabled ("no
ipv6 nd suppress-ra"). Thanks to David Lamparter for reporting and
fixing this issue.

CVE-2016-4049 A controlled crash, leading to a DoS, in the BGP MRT
route dumping code, if a prefix had too many entries to write to one
record. Thanks to Evgeny Uskov for reporting and fixing this issue.