testing/libmilter: set default pthread stack size
This patch tries to fix various crashes for applications depending on
libmilter by setting the stack size for pthreads to 8MB. The default stack size
for musl libc is set to 80KB whereas glibc has it set to 8MB. This causes
problems when a large amount of memory is allocated on the stack.

For example, opendkim allocates blocks of 64KB multiple times, which causes
libmilter (and therefore opendkim) to crash.
Maybe a stack size of 1MB or 2MB would be sufficient, but as opendkim
depends on the default glibc behavior, 8MB should be safe.

I know this patch is kind of ugly, a better solution may be to file
a request for opendkim to allocate large blocks of memory on the heap.
But as libmilter/opendkim are fairly unusable at the moment, I suggest
to apply this patch as long as these packages are in testing.

* refs: #6360