main/gcc: expand cross-compiler lib symlinks as linker scripts
This makes gcc to search the actual library from libpath. Needed
because cross-compiler's library is really in --sysroot path.
Alternative is to move the .so files to separate new target arch
specific gcc-dev package. But that'd require musl-dev or build-base
to depend on it, so this is the easy fix around it.