main/nodejs: rename to nodejs-current and move to community
Odd-numbered versions, like this one (v7), are supported by upstream
only for 9 months. When a new odd-numbered major release is cut, the
previous even-numbered major version transitions to the Long Term
Support plan (LTS).

Packages in Alpine stable must be supported for 2 years, so we should
keep only LTS version in the stable. Therefore this package is renamed
to nodejs-current and moved to the community repository. The nodejs-lts
package is going to be renamed to nodejs.

See https://github.com/nodejs/LTS#lts-schedule