testing/exim: Fix linking error on loadable modules/subpackages
Something about the gcc invocation was causing it to not link libpq or
libmysql so the exim-postgresql and exim-mysql packages were unusable at
runtime.

linux-headers build-time deb didn't seem to cause any problems, but it
issued a warning about linux/ip.h not found, and we can fix that easily
enough so we should.

Fixes #6775