testing/ghc: Split ghc into ghc and ghc-dev
While here ensure that strip --strip-unneeded is called on all
shared libraries and static archives. This reduces the overall
install sizes a skosh.

The larger win here, is stripping out the profiled debugging libraries
from the default ghc package. If someone wanted to debug now
they will need to install ghc-dev to get that RTS and debug libraries.

Realistically this only affects those users that are building
packages that need to debug. As well as future enhancements
to the ghc/haskell which will build packages with profiled libraries
as well. But this compromise for the overall install size
is preferable to users installing libraries they likely won't
need/use and overal bandwidth that consumes.