community/wait4ports: new aport
wait4ports is a small binary that waits for network peers to be ready
for connections.

I wrote wait4ports in 2015 when I could not find a small tool to
reliably poll for a list of network peers to be ready for connections.

If there is a better way in Alpine Linux, I would love to use that
instead.