main/newsbeuter: Compile without glibc
If you try to compile newsbeuter without -fpermissive, as ppc64le,
it will not build due to a wrong argument type.

It is wrong because it sets the propoer argument if you have Linux
with Glibc. Since we use MUSL, it use the wrong argument.

This patch assure that both musl and glibc uses the same argument
type.