testing/vis: upgrade to 0.3
Also add a check function while being here.

Tests are kept in a separate repository and normally checked out as a
git submodule. However, the tarball we fetch from GitHub doesn't contain
the .git directory therefore we can't use git submodules. Instead we
download the tests manually and copy them to the $builddir in check().