main/linux-vanilla: Update config for new 4.9 kernel
Current kernel 4.9 is not being built with the current config file,
since a lot of extra CONFIG is required, and it is undefined on the
current config file.

This patch's intention is to be less invasive possible, just adding
the new CONFIGs. The new CONFIG are as near as possible with x86.

This work was based on original patch from Roberto Olivera.