main/gdb: fix gdb 7.12.1 build with musl on ppc
GDB (as other applications) relies on a libc that includes Linux
headers by default, like <asm/ptrace.h>. However Musl does not rely
on these headers and instead decided to define them also. As a
consequence build brakes due to redefinitions.

That commit fixes GDB build on ppc by excluding the problematic
dependency on Linux headers and also addresses an issue that
appeared from GDB 7.11.1 to 7.12.1, when GDB started to be built
as a C++ program by default [1, 2], generating mangled and
non-mangled objects.

[1] https://goo.gl/HYL82h
[2] https://sourceware.org/ml/gdb/2016-04/msg00041.html