community/go: enable build on ppc64le.
Go package build was disabled on ppc64le because it was missing go-bootstrap dependency.
As the last version of go-bootstrap (1.4) is not available for ppc64le, I cross-compiled
it and uploaded it to: ftp://ftp.unicamp.br/pub/ppc64el/alpine/go-bootstrap/

The go-bootstrap is now installed in build-edge-ppc64le and can compile the go package.