main/libbsd: enable build on ppc64le
Libbsd includes a.out.h header to get some definitions, like struct nlist,
struct exec.
In Alpine x86_64 this header is provided by linux-headers package (linux/a.out.h),
in Ubuntu ppc64le it is provided by glibc (a.out.h), but in Alpine ppc64le it is not
available.

A workaround for this problem is to define the types that are being used
by libbsd instead of including the invalid header for ppc64le.