scripts: add support for ppc64le in vanilla configuration
To boot ppc64le we need to use grub with ieee1275 platform. Said that,
a new section to support grub ieee1275 was created.
Also needed to change xorrisofs command by grub-mkrescue, that is the one
used to create a bootable image for power and it is also a wrapper to xorriso.