community/elasticsearch: update to 5.4.2 / add bash depends
* also fixes plugin path:

ES expects to find plugins in the same path as the libs. Looking at
the java command line in htop ES sets path.home a 2nd time for the
libs & expects the plugins to also be in this path (which is why
there is no configuration for the plugin path in elasticsearch.yml).