main/llvm4: rename lit subpkg to -test-utils, move FileCheck and not here
It seems that overriding pkgver for a subpackage really wasn't
a good idea. I still don't know what the heck happened on the build
servers, but for some unknown reason lit-*.apk was not uploaded to
mirrors. Also builders rebuilt llvm4 twice, even when the previous
build succeeded.

Anyway, I'm even more convinced that it was just a bad idea. Upstream
does not actually release lit, they release only llvm with bundled
lit. It has some version number in setup.py, but what if they let it
unchanged between two llvm releases, despite making some changes in
the lit's codebase?

It's true that it's wrong to call it lit-4.0.0 when its installed as
a python module with version number 0.6.0. So I kinda workarounded it
by renaming this subpackage to llvm4-test-utils with
provides="lit-0.6.0-rX". :) And to not leave lit here alone, I've
also moved FileCheck and not utils here, 'cause these are testing
utils too and lit needs them.