community/cabal: Update APKBUILD with more stable dependencies
Instead of using the regular .cabal uri's, which can be changed
to allow old packages to build against newer ghc's or to fix issues
with old package authors not updating their .cabal files, use
the specific revision of the cabal files we used for bootstrapping
cabal itself.

Also add -$pkgver-$pkgrel to .cabal files from hackage, in the prepare()
function move them to package.cabal so that bootstrap.sh works.