non-free/intel-ucode: Update to 20170511
In the context of the recent Skylake/Kaby Lake hyper-threading bug [1], an
update of the intel microcode seems to be quite relevant. The resulting
package will contain "/boot/intel-ucode.img", which needs to be added as
first initial ram drive. This is achieved in extlinux by chaining the line

  INITRD=linux-$tag

to

  INITRD=intel-ucode.img,linux-$tag

See [2] for more details. If this packet is accepted, changing
/sbin/update-extlinux to check for "/boot/init-ucode.img" and adding it
as first INITRD on detection seems to be a good idea.

All credit goes to Thomas Bächler [3], from whom I shamelessly stole the build
recipe.

[1] https://lists.debian.org/debian-devel/2017/06/msg00308.html
[2] https://wiki.archlinux.org/index.php/Microcode#Syslinux
[3] https://git.archlinux.org/svntogit/packages.git/tree/trunk/PKGBUILD?h=packages/intel-ucode