community/evince: include all commits from gnome-3-24 git branch
Upstream doesn't seem to do minor releases and instead simply pushes
security and reliability fixes to the git branch associated with the
latest releases. Therefore it might make sense to build evince directly
from the git repository in the future, however, for now it should be
sufficient to include the patches as git-format(1) patches.