community/ghc: upgrade to 8.2.1
There is only one package that depends on ghc right now, and that is
community/cabal. The move to community and the subsequent cabal file
bounds that added made the older cabal-install build with 8.2.1 so
no real reason to hold off upgrading ghc.

Removed the haskell-* provides as that was originally a plan to follow
what arch was doing for their haskell setup. In lieu of that not
providing value removing them entirely.

The ghc test suite was updated to use python 3 from python 2.

Finally really late in the release candidate cycle a change to ld
detection necessitated the need to pass --disable-ld-override to the
configure of ghc.

Without that in place there appears to be some shenanigans going on
in exported symbols.

Ghc upstream ticket:
https://ghc.haskell.org/trac/ghc/ticket/13958

Offending commmit:
https://git.haskell.org/ghc.git/commitdiff/2785ef0e31a123400da950ffafebe6cb1ce3f4eb