community/ghc: upgrade to 8.4.2
There is only one package that depends on ghc right now, and that is
community/cabal.

Removed armhf support for now as there appear to be overall issues with
that architecture across libc's. Additionally, for 8.6 to build we will
need to update to at least 8.2. As such, dropping support for armhf
makes the most sense for right now. I'll continue working with upstream
to fix it.

Removed the haskell-* provides as that was originally a plan to follow
what arch was doing for their haskell setup. In lieu of that not
providing value removing them entirely.

The ghc test suite was updated to use python 3 from python 2.

Finally really late in the release candidate cycle a change to ld
detection necessitated the need to pass --disable-ld-override to the
configure of ghc.

Without that in place there appears to be some shenanigans going on
in exported symbols.

Ghc upstream ticket:
https://ghc.haskell.org/trac/ghc/ticket/13958

Offending commmit:
https://git.haskell.org/ghc.git/commitdiff/2785ef0e31a123400da950ffafebe6cb1ce3f4eb