testing/texmf-dist: new aport
TeX Live texmf core distribution
http://tug.org/texlive/

The main package "texmf-dist" provides the core texmf-dist
distribution. This allows texlive to work, as long as only the
most basic LaTeX packages are used.

The sub-package "texmf-dist-most" provides most of the texlive
packages. It depends on texmf-dist, as it only adds packages not
included in texmf-dist.

This is the least afford way to get LaTeX working in Alpine Linux.
Ideally, the texlive packages should be divided in multiple
separate packages to reduce required network bandwidth on updates.
As always with any LaTeX code and LaTeX related code, the trigger
is a piece of cargo cult programming that happened to work...

ref #4969