main/vlan: check if subinterface already exists
The vlan preup script tried to create the subinterface even if it
already exists. It does not work on dualstack (ipv4 + ipv6)
configurations, and on subinterfaces with multiple addresses setup.

Now the vlan and mvlan scripts check if the interface does not
already exist, and only if it does not, try to create it.