main/musl: fix possible buffer overrun in __posix_spawnx().
This can allow programs to crash when $PATH contains elements that are larger than 1024 bytes.
Notably, PATH_MAX is larger than 1024 bytes, so this can cause environments with very long, but
valid path elements in $PATH to have program crashes when they spawn new processes.

No CVE for this one at the time of writing, but this seems like a probable security bug.