community/thttpd: fix serving of .gz and .Z files
thttpd erroneously treats these files as a compressed transfer
encoding rather than as a content type. this causes conforming http
clients to strip the compression and save a decompressed version when
downloading. since this was historically a common httpd bug, some web
browsers work around the problem by detecting the .gz extension and
ignoring the server's reported transfer encoding, but others,
including wget 1.19.2, save a decompressed file, breaking file
integrity checking (based on a hash or signature of the original
compressed file) and breaking scripts which pass the -z option to tar
when extracting.

add a patch which removes thttpd's support for extension-based content
transfer encodings, and adds the missing mime types for gzip and
compress. the patch has been written to be minimally invasive to the
program logic, and thus leaves a for loop that breaks on the first
iteration rather than rewriting it.