main/python3: increase thread stack size
we segfault before we hit the sys.getrecurslimit(), which is bad. Some
testing showed that to be able to reach the default
sys.getrecurslimit(), which is 1000, we need at least a 453k stack on
x86_64. We set it to 1MB so we have a bit extra.

It is also worth to note that upstream set stacksize to 4MB on freebsd
and 5MB on OSX.

ref #8134
https://github.com/docker-library/python/issues/211