main/py-attrs: new aport
https://attrs.readthedocs.org
Python Classes Without Boilerplate

main/py-automat: new aport

https://github.com/glyph/automat
Self-service finite-state machines for the programmer on the go.

main/py-constantly: new aport

https://github.com/twisted/constantly
Symbolic Constants in Python

Added remaining py-twisted dependencies, for fixing #7154 and #7157. There is still the decision to make if we should also bring py-service_identity to main, an optional dependency that yields a warning when runnning either deluge or deluge-web. Unfortunately, it also has dependencies in the community repo.