main/augeas: fix test
problem with test/test-preserve.sh was that it compared the group column
from out put of ls. If the group name was longer than 8 the test would
fail because busybox' ls truncate the groupname.

We solve this by using stat -c %G instead, which does not truncate the
group name.