abuild: Change -f (force) to not ignore checksums/init scripts
The force flag used to skip the following functions, without any
documentation in the help (-h) output:

  * verify (checksum verification)
  * initdcheck (check if the init scripts are openrc scripts)
  * check_arch (check if the target architecture is in "arch=")
  * check_libc (check if the target libc is masked in the options)

This was counter-intuitive and could even be dangerous (when one relies
on the checksum verification to prevent man-in-the-middle attacks, but
always uses the -f flag).

With this commit, it only skips check_arch and check_libc besides the
package up to date check and the help output mentions this.