main/wpa_supplicant: refactor wpa_cli.sh
udhcpc doesn't have link beat detection and the purposes of this script
is making udhcpc acquire a new lease when connecting to an access point.

The previous version of this script achieved that goal by running
ifdown/ifup which is pretty ugly and never worked properly (at least
with my setup). A better solution for acquiring a DHCP lease when
connecting to a new AP is sending a SIGUSR1 signal to the udhcpc running
on the affected interfaces, this solution is implemented in this commit.

Together with the wpa_cli OpenRC service added in the previous commit
this makes using udhcpc on laptops very convenient.

See also: cc8f61f6330f49f1e2522e6fbed88e84e47190c3