main/libgssglue: depend on heimdal-dev and heimdal-libs
This is necessary because the `/etc/gssapi_mech.conf` configuration file
shipped by this package refers to `/usr/lib/libgssapi.so` which is
shipped by heimdal-dev and is a symlink to libgssapi.so.3.0.0 which is
in turn shipped by heimdal-libs.

Packages like `main/rdesktop` parse `/etc/gssapi_mech.conf` and load the
shared libraries referred by it which causes `main/rdesktop` to emit the
following warning:

can't open /usr/lib/libgssapi.so: Error loading shared library /usr/lib/libgssapi.so: No such file or directory

This change fixes this.