main/sic: apply two official patches
The first patch is lineprint.patch.
This patch allow to store and reprint the current inserting line.
This is done by appending a suffix character ('\') to the input.
The line is (obviously) not sent.
This is useful, for example, when receiving data from the server in the middle of a sentence you’re writing.

The second patch is hidecommand.patch.
This patch lets you hide IRC messages such as JOIN/QUIT.
You can use the sic command ‘h’ to specify what messages are hidden.
Example: :h JOIN,QUIT
Now all JOIN and QUIT messages will not be displayed.
The delimiter is unimportant (use what you like or nothing at all)
Use the command without arguments to stop hiding any messages.