main/sdl2: fix DirectFB renderer regression
The update to 2.0.8 broke the build of the DirectFB renderer such that
it cannot be used anymore, see
https://bugzilla.libsdl.org/show_bug.cgi?id=4123. This fixes the
regression by re-inserting the affected line in the configure script.

Having a working DirectFB renderer in SDL2 is crucial for postmarketOS
as it is used for unlocking full disk encryption and displaying the
charging animation on almost all devices, see also postmarketOS/osk-sdl#80.