main/json-c0.12: temporary aport
we need a libjson-c.so.2 til we fully migrated to libjson-c.so.4 (from
json-c 0.13) This is because mkinitfs is installed on builders so they
can make relases and mkinitfs depends on cryptsetup which depends on
json-c. This means that we have a dependency installed for
so:libsjon.so.2 which holds back json-c-dev to install version 0.13.

We solve this by adding a temporary json-c0.12 which can provide
so:libjson-c.so.2.