community/py3-aiohttp-cors-gns3: new aport
This has been introduced due to a long required compatibility
for GNS3.
In alpine we are upgrading to the latest version (which often breaks ABI), while GNS3 sticks to older
versions.
This new aport introduce a layer of compatibility, following what also Archlinux does.
It is pushed in community directly in order to be a dependency of GNS3, though has been tested.