testing/rakudo: disable on x86
build fails with:
===SORRY!===
The nqp-m binary is too old

To automatically clone (git) and build a copy of NQP 2018.04,
try re-running Configure.pl with the '--gen-moar' option.
Or, use '--prefix=' to explicitly specify the path where the NQP and
MoarVM
executables can be found that are use to build Rakudo.