main/openssh: fix flavour split function, add kerberos
The support for different flavours of openssh-server packages
was broken, effectively only supporting a single additional
flavour. The changes now support multiple flavours as
subpackages and also use the '+' sign in the flavour name
as a separator for indicating multiple packages which need
to be enabled in the openssh build. E.g. the "kerberos5+pam"
flavour enables both kerberos and pam support in a single
subpackage named openssh-server-kerberos5+pam.