main/busybox: properly fix wget https support
fix busybox wget https support by using an external ssl_client helper
for https.

Disable the use of external openssl. This was fixed to check
certificates as a temporary solution. openssl can not produce any useful
error messages on certificate errors. It is big. So we simply disable
its use.

For dynamic busybox we disable the internal ssl_client and the internal
(broken) tls code, and build our own ssl_client which properly verifies
the certificates.

For the static busybox we enable the internal ssl_client and tls code,
but we only allow its use with --no-check-certificates. This is so we
still can fetch things from https in an emergency situation.

We auto-install ssl_client if both libssl and busybox are installed. This
is to keep backwards compatibility.