main/busybox: properly fix wget https support
fix busybox wget https support by using an external ssl_client helper
for https.

Disable the use of external openssl. This was fixed to check
certificates as a temporary solution. openssl can not produce any useful
error messages on certificate errors. It is big. So we simply disable
its use.

We auto-install ssl_client if both libssl and busybox are installed. This
is to keep backwards compatibility.