main/alpine-baselayout: revert 'sysctl security changes'
Lack of agreement on proper values for these settings indicates
that we should be following the Linux kernel's defaults
rather than enforcing our own.

Multiple mistakes in the original implementation reinforce the
justification for this principle.

We can selectively add back any sysctl settings that have a clear
and compelling need to be overriden.