Use the proper getty for login on ppc64le
On ppc64le when net booting Alpine we will not get a login prompt from
Petitboot. Despite hvc0 being the default TTY, we need to make sure
getty uses the correct console for a login prompt. This patch will set
TTY as /dev/console and the TERMTYPE as vt100 while using a safer BAUD
value in order to work on varying ppc64le hardware.