main/openssl1.0: rebuild
I've renamed the aport in commit 2fdc8bcc6549290b131675ef42d52243c37f3879
and so resetted pkgrel to 0. The problem is that the name of subpackages
libcrypto1.0 and libssl1.0 (intentionally) remained unchanged - they
already contain the version suffix. So for them the pkgrel was
effectively decreased, which I did not realize. :(

Fixes #9069 (https://bugs.alpinelinux.org/issues/9069)