Ignore poorly designed custom python unit test issues
The issue seems to be well summarized here:

https://stackoverflow.com/questions/6380057/python-binding-socket-address-already-in-use/18858817#18858817

but in any case, it looks as though python processes are left hanging because the
client side does not terminate the connection.

Since we're using port 9090 by default, and another process is still listening on
that port, new processes receive EADDRINUSE.

This particularly effects TestServer.py . If it were immediately obvious how to fix
this brokenness I would fix it. Some people suggest using addCleanup() in another
python unit testing framework, but the thrift people have opted to implement their
own python unit testing framework, which is obviously sub-par.

I've implemented the same scenario (as described by strace) in C and the issue does
not occur.